nrf51_sdk_v6_1_0_b2ec2e6.zip

+++ nrf51422/Board/nrf6310/s210/ant_bicycle_pwr/bicycle_power_only_sensor/main.c
@@ -47,7 +47,7 @@ static void softdevice_assert_callback(uint32_t pc, uint16_t line_num, const uin
 {
 #if defined(TRACE_UART)    
     printf("ASSERT-softdevice_assert_callback\n");
-    printf("PC: %#x\n", pc);
+    printf("PC: %#x\n", (unsigned int)pc);
     printf("File name: %s\n", (const char*)p_file_name);
     printf("Line number: %d\n", line_num);
 #endif // defined(TRACE_UART)     
@@ -69,9 +69,9 @@ void app_error_handler(uint32_t error_code, uint32_t line_num, const uint8_t * p
 {
 #if defined(TRACE_UART)    
     printf("ASSERT-app_error_handler\n");
-    printf("Error code: %u\n", error_code);
+    printf("Error code: %u\n", (unsigned int)error_code);
     printf("File name:%s\n", (const char*)p_file_name);
-    printf("Line number: %d\n", line_num);
+    printf("Line number: %d\n", (unsigned int)line_num);
 #endif // defined(TRACE_UART)         
   
     for (;;)
