+++ nrf51422/Board/pca10003/s310/ant_hrm/hrm_rx/main_hrm_rx.c
@@ -53,7 +53,7 @@ void softdevice_assert_callback(uint32_t pc, uint16_t line_num, const uint8_t * 
 {
 #if defined(TRACE_UART)
     printf("ASSERT-softdevice_assert_callback\n");
-    printf("PC: %#x\n", pc);
+    printf("PC: %#x\n", (unsigned int)pc);
     printf("File name: %s\n", (const char*)p_file_name);
     printf("Line number: %u\n", line_num);
 #endif // TRACE_UART   
@@ -75,9 +75,9 @@ void app_error_handler(uint32_t error_code, uint32_t line_num, const uint8_t * p
 {
 #if defined(TRACE_UART)
     printf("ASSERT-app_error_handler\n");
-    printf("Error code: %u\n", error_code);
+    printf("Error code: %u\n", (unsigned int)error_code);
     printf("File name: %s\n", (const char*)p_file_name);
-    printf("Line number: %u\n", line_num);
+    printf("Line number: %u\n", (unsigned int)line_num);
 #endif // TRACE_UART   
 
     for (;;)
